+++ lib/utils/common.dart
@@ -5,12 +5,13 @@ import 'package:flutter/services.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:revelation/app_router.dart';
 import 'package:logger/logger.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:xml/xml.dart';
-import 'dependent.dart';
+import '../app_router.dart';
+import '../models/topic_info.dart';
 import '../models/library_info.dart';
+import 'dependent.dart';
 
 final log = Logger();
 
@@ -26,6 +27,15 @@ bool isWeb() {
 
 TargetPlatform getPlatform() {
   return defaultTargetPlatform;
+}
+
+String locLinks(BuildContext context, String key) {
+  final localizations = AppLocalizations.of(context);
+  final Map<String, String> translations = {
+    "topic_1_name": localizations!.topic_1_name,
+    "topic_1_description": localizations.topic_1_description,
+  };
+  return translations[key] ?? key;
 }
 
 String getSystemLanguage() {
@@ -82,6 +92,43 @@ Future<List<LibraryInfo>> parseLibraries(
     }
 
     return libraries;
+  } on XmlException {
+    rethrow;
+  } on PlatformException {
+    rethrow;
+  } catch (e) {
+    throw Exception('Unknown error: $e');
+  }
+}
+
+Future<List<TopicInfo>> parseTopics(AssetBundle bundle, String xmlPath) async {
+  try {
+    final xmlString = await bundle.loadString(xmlPath);
+    final document = XmlDocument.parse(xmlString);
+    final topics = <TopicInfo>[];
+
+    for (var element in document.findAllElements('topic')) {
+      final name = element.getElement('name')?.innerText;
+      final idIcon = element.getElement('idIcon')?.innerText;
+      final description = element.getElement('description')?.innerText;
+      final route = element.getElement('route')?.innerText;
+
+      if (name == null ||
+          idIcon == null ||
+          description == null ||
+          route == null) {
+        throw Exception('Missing required tags in topic element');
+      }
+
+      topics.add(TopicInfo(
+        name: name,
+        idIcon: idIcon,
+        description: description,
+        route: route,
+      ));
+    }
+
+    return topics;
   } on XmlException {
     rethrow;
   } on PlatformException {