+++ CHANGELOG.md
@@ -9,8 +9,10 @@ All notable changes to this application will be documented here.
 - Support languages: English, Ukrainian, Russian;
 - Support platforms: Android, iOS, Windows, macOS, linux, web;
 - Main menu (left side);
-- About screen (Common info, Contact link, Acknowledgments and 3rd party licenses, Changelog);
+- Main screen (List of topics);
 - Settings screen (Change language)
+- About screen (Common info, Contact link, Acknowledgments and 3rd party licenses, Changelog);
+
 
 ### Changed